update user to create model properly

+++ core/controllers/user_api.py
@@ -1,6 +1,8 @@
 from django.shortcuts import get_object_or_404
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
+from rest_framework.decorators import permission_classes
+from rest_framework.permissions import IsAdminUser, IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
@@ -29,7 +31,25 @@ class UserGetAllCreateApiView(APIView):
         user_data = self.create_serializer(data=self.request.data)
         user_data.is_valid(raise_exception=True)
 
-        new_user = self.repo(self.model_class).create(**user_data.validated_data)
+        username = user_data.validated_data.get("username")
+        password = user_data.validated_data.get("password")
+
+        if not username:
+            return Response(status=400, data={"message": "Username is None"})
+        if not password:
+            return Response(status=400, data={"message": "Password is None"})
+
+        if self.model_class.objects.filter(username=username).exists():
+            return Response(
+                status=400,
+                data={
+                    "message": f"Username with this {username} username already exists"
+                },
+            )
+
+        new_user = self.model_class.objects.create_user(
+            username=username, password=password, **user_data.validated_data
+        )
         return Response(self.detail_serializer(new_user).data)
 
     @swagger_auto_schema(
@@ -39,6 +59,7 @@ class UserGetAllCreateApiView(APIView):
             )
         }
     )
+    @permission_classes([IsAdminUser])
     def get(self, request):
         queryset = self.repo(self.model_class).get_all()
         serializer = self.detail_serializer(queryset, many=True)
@@ -55,8 +76,14 @@ class UserGetDeleteUpdateApiView(APIView):
     @swagger_auto_schema(
         responses={"200": openapi.Response("response description", detail_serializer)}
     )
+    @permission_classes([IsAuthenticated])
     def get(self, request, pk=None):
         queryset = self.repo(self.model_class)
+        request_user = self.request.user
+
+        if request_user.id != pk:
+            return Response(status=403)
+
         user = get_object_or_404(queryset, pk=pk)
 
         serializer = self.detail_serializer(user)
@@ -68,7 +95,13 @@ class UserGetDeleteUpdateApiView(APIView):
             "200": openapi.Response("detail user serializer", update_serializer)
         },
     )
+    @permission_classes([IsAuthenticated])
     def put(self, request, pk=None):
+        request_user = self.request.user
+
+        if request_user.id != pk:
+            return Response(status=403)
+
         repo = self.repo(self.model_class)
         old_user = repo.get(pk)
         data = self.update_serializer(old_user, data=self.request.data)
@@ -78,7 +111,13 @@ class UserGetDeleteUpdateApiView(APIView):
 
         return Response(data.data)
 
+    @permission_classes([IsAuthenticated])
     def delete(self, request, pk: str = None):
+        request_user = self.request.user
+
+        if request_user.id != pk:
+            return Response(status=403)
+
         try:
             repo = self.repo(self.model_class)
             repo.delete(pk)
